+++ CHANGELOG.rst
@@ -22,6 +22,7 @@ Unreleased
 
 **Other changes**
 
+- Optimized the :meth:`quantcore.matrix.SparseMatrix.matvec` and :meth:`quantcore.matrix.SparseMatrix.tranpose_matvec` for when ``rows`` and ``cols`` are None.
 - Implemented :func:`CategoricalMatrix.__rmul__`
 - Reorganizing the documentation and updating the text to match the current API.
 - Enable indexing the rows of a ``CategoricalMatrix``. Previously :func:`CategoricalMatrix.__getitem__` only supported column indexing.